no errors on startup when included in Gemfile
