+++ .gitignore
@@ -7,3 +7,4 @@
 tmp
 nbproject
 *.swp
+pkg
