--- app/controllers/addresses_controller_decorator.rb
@@ -1,3 +1,0 @@
-AddressesController.class_eval do
-  helper :checkout
-end
