+++ app/controllers/spree/admin/orders_controller_decorator.rb
@@ -1,4 +1,4 @@
-Admin::OrdersController.class_eval do
+Spree::Admin::OrdersController.class_eval do
   def courier_instruction
     @order = Order.find_by_number(params[:id])
     render :layout => false
