+++ app/controllers/spree/admin/products_controller_decorator.rb
@@ -1,5 +1,5 @@
 # coding: utf-8
-Admin::ProductsController.class_eval do
+Spree::Admin::ProductsController.class_eval do
   require 'nokogiri'
   require 'open-uri'
