+++ app/controllers/spree/checkout_controller_decorator.rb
@@ -1,8 +1,8 @@
-CheckoutController.class_eval do
+Spree::CheckoutController.class_eval do
   private
 
   def after_payment
-    if @order.shipping_method.calculator.is_a?(Calculator::CashOnDelivery)
+    if @order.shipping_method.calculator.is_a?(Spree::Calculator::CashOnDelivery)
       @order.payments.build(:payment_method => PaymentMethod.internal, :amount => @order.total)
       @order.confirm_without_payment!
     end
