+++ app/models/spree/calculator/cash_on_delivery.rb
@@ -1,4 +1,4 @@
-class Calculator::CashOnDelivery < Calculator
+class Spree::Calculator::CashOnDelivery < Spree::Calculator
   def self.description
     I18n.t("cash_on_delivery")
   end
