+++ app/models/spree/payment_method/juridical_invoice.rb
@@ -1,4 +1,4 @@
-class PaymentMethod::JuridicalInvoice < PaymentMethod::Check
+class Spree::PaymentMethod::JuridicalInvoice < Spree::PaymentMethod::Check
   def desc
     "<p>#{I18n.t('juridical_invoice')}</p>".html_safe
   end
