+++ app/models/spree/user_decorator.rb
@@ -1,4 +1,4 @@
-User.class_eval do
+Spree::User.class_eval do
   validates :recipient, :juridical_address, :inn, :recipient_bank,
             :settlement_acc, :corr, :bik, :presence => true, :if => :is_juridical?
   validates :inn, :numericality => true, :length => { :is => 10 }, :if => :is_juridical?
