+++ db/migrate/20090827180351_add_secondname_to_addresses.rb
@@ -1,9 +1,5 @@
 class AddSecondnameToAddresses < ActiveRecord::Migration
-    def self.up
-      add_column :addresses, :secondname, :string
-    end
-
-    def self.down
-      remove_column :addresses, :secondname
-    end
+  def change
+    add_column :spree_addresses, :secondname, :string
+  end
 end
