--- db/migrate/20110818091541_change_user_juridical_fields.rb
@@ -1,11 +1,0 @@
-class ChangeUserJuridicalFields < ActiveRecord::Migration
-  def self.up
-    change_column :users, :inn, :string
-    change_column :users, :kpp, :string
-  end
-
-  def self.down
-    change_column :users, :kpp, :integer
-    change_column :users, :inn, :integer
-  end
-end
